Make SearchView viewable
Add search suggestions
